refactor: moved package to  prizeprovider and .impl
