--- src/main/java/cc/oakk/lotto/model/LottoPrizeProvider.java
@@ -1,5 +1,0 @@
-package cc.oakk.lotto.model;
-
-public interface LottoPrizeProvider<T> {
-    LottoPrize<T> getPrizeByRank(Rank rank);
-}
